:octocat: Fixed Kommands and Updating PlayerFace
RCONBase had its connect() function update to try/exception in order to
return a True or False for the connection attempt. Kommands had the
PlayerStruc overhauled and the getplayerlist() changed to have its
fields converted into an appropriate type. (Position is now a list - X =
0, Y = 1,  Z = 2).

Limiter works... halfway. Detection works fine, but action should be
user-defined, so i'll just provide a simple example for when a detection
is raised.