:honeybee: Bkthread Hotfix Implementation
The problem was detected on the 10th, but due to unfamiliarity with
django, I didn't want to release it under the assumption that this would
be safe for a web framework.

The bkthread allows the BF2CC class to update a single list containing
either bf2cc si or pl so other functions and properties can call back
new info in a timely fashion. You can adjust the timer.sleep to whatever
you want, and the thread is killed as soon as the main application is
killed.

For Django and other web uses, this isn't recommended, and other
services like Celery are optimal for queuing tasks.